requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
joblib==1.4.2
lxml==5.3.1
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
python-docx==1.1.2
python-pptx==1.0.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.0
urllib3==2.3.0
XlsxWriter==3.2.2
